Working on search text field
